refactor send email functionality, #319
