+++ lib/email.js
@@ -1,9 +1,15 @@
+'use strict';
+var google       = require('googleapis');
+var OAuth2       = google.auth.OAuth2;
+var oauth2Client = new OAuth2(process.env.GOOGLE_CLIENT_ID, process.env.GOOGLE_CLIENT_SECRET);
+
 var Joi = require('joi');
-var sendEmail = require('./helpers/sendemail_gmail.js');
+// var sendEmail = require('./helpers/sendemail_gmail.js');
+const sendEmail = require('./helpers/email/send_email');
 var candidatesEmailId = require('./helpers/candidates_email_id.js');
 var client = require('./es.js');
-var capitalizeFirstLetter = require('./helpers/capitalize_first_letter.js');
-var generateSignature = require('./helpers/generate_signature');
+var capitalizeFirstLetter = require('./helpers/email/capitalize_first_letter.js');
+var generateSignature = require('./helpers/email/generate_signature');
 
 exports.register = function (server, options, next) {
 
@@ -12,12 +18,13 @@ exports.register = function (server, options, next) {
     method: 'POST',
     path: '/email',
     config: {
-      description: 'create email',
+      description: 'create email from the selected profiles',
       auth: {
         mode: 'try',
         strategy: 'jwt'
       },
       handler: function (request, reply) {
+
         if (!request.auth.isAuthenticated) {
           return reply.redirect('/login');
         }
@@ -25,6 +32,7 @@ exports.register = function (server, options, next) {
         var schema = {
           email: Joi.required()
         }
+
         var validationObject = Joi.validate(request.payload, schema, { abortEarly: false, allowUnknown: true });
 
         if (!validationObject.error) {
@@ -51,6 +59,7 @@ exports.register = function (server, options, next) {
             type: process.env.ES_TYPE_USERS,
             id: request.auth.credentials.id
           }, function (err, response) {
+
             var user = response._source;
             var userObject = {
               fn: user.names.fullname,
@@ -74,7 +83,7 @@ exports.register = function (server, options, next) {
     method: 'POST',
     path: '/sendemail',
     config: {
-      description: 'send email',
+      description: 'send the email to the profiles',
       auth: {
         mode: 'try',
         strategy: 'jwt'
@@ -94,6 +103,7 @@ exports.register = function (server, options, next) {
         var candidateFirstName = request.payload.candidateFN;
         var allEmails = [];
 
+        //sent to just one profile -> create the array
         if (typeof emailList === 'string') {
           allEmails.push(emailList);
         } else {
@@ -113,14 +123,19 @@ exports.register = function (server, options, next) {
           candidatesFirstNames = candidateFirstName;
         }
 
+        //data for the user signature
         var user = {};
         user.fn = request.payload.fn;
         user.role = request.payload.role;
         user.office = request.payload.office;
         user.mobile = request.payload.mobile;
         user.linkedin = request.payload.linkedin;
+
+        //create an array of profiles for the email [{ email: emailOfTheProfile, id: idOfTheProfile, firstName: firstNameProfile },...]
         var arrCandidateObj = candidatesEmailId(allEmails, allCandidatesIDs, candidatesFirstNames);
+        //profile of the user who sent the emails
         var profile = request.auth.credentials;
+        //create common data of each emails
         sentEmail.subject = request.payload.subject;
         sentEmail.message = request.payload.message;
         sentEmail.senderName = JSON.parse(process.env.MAP_ID_USER)[request.auth.credentials.id];
@@ -130,18 +145,21 @@ exports.register = function (server, options, next) {
         var currentDate = new Date();
         sentEmail.sentAt = currentDate.getDate() + '-' + ("0" + (currentDate.getMonth() + 1)).slice(-2) + '-' + currentDate.getFullYear();
 
+        //authenticate to use gmail api
+
+        oauth2Client.setCredentials(profile.tokens);
+        //sent email for each profiles
         var counter = 0;
         arrCandidateObj.forEach(function (obj) {
 
+          sendEmail(obj, sentEmail, oauth2Client, user, function (err, response) {
 
-          sendEmail(obj, sentEmail, profile, user, function (err, response) {
             if (err) {
               console.log(err);
-              //return reply('something went wrong');
               return reply.redirect('/login');
-
             }
 
+            //if the email is sent, updat the emails property of the profiles by adding the new email to the list
             if (response.labelIds.indexOf('SENT') !== -1) {
 
               client.get({
@@ -151,10 +169,13 @@ exports.register = function (server, options, next) {
                 _source: ['emails']
 
               }, function (errElastic, responseElastic) {
-                if (errElastic) {
-                  next(errElastic);
-                }
+
+                  if (errElastic) {
+                    next(errElastic);
+                  }
+
                   var currentEmail = responseElastic._source.emails;
+
                   if (currentEmail === undefined) {
                     currentEmail = [];
                   }
@@ -188,7 +209,9 @@ exports.register = function (server, options, next) {
             }
 
             counter++;
+            //redirect to the profile or the home page
             if (counter === allEmails.length) {
+
               if(request.payload.pathUrl){
                 return reply.redirect(request.payload.pathUrl);
               }
