+++ lib/helpers/email/capitalize_first_letter.js
