+++ lib/helpers/email/generate_signature.js
